app/requirements.txt
django==3.1.1
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytest-django==3.9.0
pytest==5.4.1